feat: 新增动态、评论、互动、通知功能，统一用户ID类型，修复数据库形式和model不匹配

+++ api/routers/v1/user/profile/profile.go
@@ -63,7 +63,7 @@ func UpdatePartialHandler(ctx fiber.Ctx, servants *server.Servants) error {
 	return updateProfile(ctx, servants, uid, updates)
 }
 
-func updateProfile(ctx fiber.Ctx, servants *server.Servants, uid uint64, updates map[string]any) error {
+func updateProfile(ctx fiber.Ctx, servants *server.Servants, uid uint32, updates map[string]any) error {
 	user, err := servants.UserServant.UpdateProfile(uid, updates)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -153,7 +153,7 @@ func validNickname(value string) bool {
 	return value != "" && len(value) <= maxNicknameLen
 }
 
-func validRegion(value int) bool {
+func validRegion(value uint16) bool {
 	return value > 0
 }
