+++ dao/user/update.go
@@ -2,7 +2,7 @@ package user
 
 import modeluser "miaoverse/model/dao/user"
 
-func (d *UserDAO) UpdateProfile(userID uint64, updates map[string]any) (*modeluser.User, error) {
+func (d *UserDAO) UpdateProfile(userID uint32, updates map[string]any) (*modeluser.User, error) {
 	if err := d.DB.Model(&modeluser.User{}).Where("id = ?", userID).Updates(updates).Error; err != nil {
 		return nil, err
 	}
